Refactored the package structure
