+++ java/code/src/com/suse/manager/hub/DefaultIssInternalClient.java
@@ -9,11 +9,13 @@
  * http://www.gnu.org/licenses/old-licenses/gpl-2.0.txt.
  */
 
-package com.suse.manager.iss;
+package com.suse.manager.hub;
 
 import com.redhat.rhn.common.util.http.HttpClientAdapter;
-import com.redhat.rhn.domain.iss.IssRole;
 
+import com.suse.manager.model.hub.IssRole;
+import com.suse.manager.model.hub.RegisterJson;
+import com.suse.manager.model.hub.SCCCredentialsJson;
 import com.suse.manager.webui.controllers.ECMAScriptDateAdapter;
 
 import com.google.gson.Gson;
