+++ java/code/src/com/suse/manager/model/hub/IssHub.java
@@ -12,7 +12,6 @@ package com.suse.manager.model.hub;
 
 import com.redhat.rhn.domain.BaseDomainHelper;
 import com.redhat.rhn.domain.credentials.SCCCredentials;
-import com.redhat.rhn.domain.iss.IssRole;
 
 import java.util.Objects;
