+++ java/code/src/com/suse/manager/model/hub/SCCCredentialsJson.java
@@ -9,7 +9,7 @@
  * http://www.gnu.org/licenses/old-licenses/gpl-2.0.txt.
  */
 
-package com.suse.manager.iss;
+package com.suse.manager.model.hub;
 
 import java.util.Objects;
