+++ java/code/src/com/suse/manager/xmlrpc/iss/IssHandler.java
@@ -11,7 +11,6 @@
 
 package com.suse.manager.xmlrpc.iss;
 
-import com.redhat.rhn.domain.iss.IssRole;
 import com.redhat.rhn.domain.user.User;
 import com.redhat.rhn.frontend.xmlrpc.BaseHandler;
 import com.redhat.rhn.frontend.xmlrpc.InvalidParameterException;
@@ -20,7 +19,8 @@ import com.redhat.rhn.frontend.xmlrpc.TokenAlreadyExistsException;
 import com.redhat.rhn.frontend.xmlrpc.TokenCreationException;
 import com.redhat.rhn.frontend.xmlrpc.TokenExchangeFailedException;
 
-import com.suse.manager.model.hub.HubManager;
+import com.suse.manager.hub.HubManager;
+import com.suse.manager.model.hub.IssRole;
 import com.suse.manager.webui.utils.token.TokenBuildingException;
 import com.suse.manager.webui.utils.token.TokenException;
 import com.suse.manager.webui.utils.token.TokenParsingException;
